17/11/2024 12:14

+++ CityGusa-api/src/main/java/com/citygusa/com/citygusaapi/Dto/ControleOperacionalDto.java
@@ -30,8 +30,8 @@ public class ControleOperacionalDto {
     private Integer gusaKg;
 
     private Integer acumuladoCarga;
-    private Double mediaHoraCarga;
-    private BigDecimal rt;
+    private BigDecimal mediaHoraCarga;
+    private Integer rt;
 
 
     public ControleOperacionalDto(ControleOperacionalEntity entity) {
